+++ sigul-pesign-bridge/src/lib.rs
@@ -11,6 +11,7 @@ mod service;
 
 use std::path::PathBuf;
 
+use anyhow::Context as AnyhowContext;
 #[doc(hidden)]
 pub use service::listen;
 
@@ -19,16 +20,12 @@ pub use service::listen;
 #[doc(hidden)]
 pub struct Context {
     pub(crate) runtime_directory: PathBuf,
-    pub(crate) credentials_directory: PathBuf,
     pub(crate) config: config::Config,
+    pub(crate) sigul_client: siguldry::client::Client,
 }
 
 impl Context {
-    pub fn new(
-        config: config::Config,
-        runtime_directory: PathBuf,
-        credentials_directory: PathBuf,
-    ) -> anyhow::Result<Self> {
+    pub fn new(config: config::Config, runtime_directory: PathBuf) -> anyhow::Result<Self> {
         // if multiple runtime directories were provided, we don't know which to use so panic for now.
         if runtime_directory
             .to_str()
@@ -42,10 +39,25 @@ impl Context {
             ));
         }
 
+        let tls_config = siguldry::client::TlsConfig::new(
+            &config.sigul.client_certificate,
+            &config.sigul.private_key,
+            None, // The expectation is the key is encrypted via systemd
+            &config.sigul.ca_certificate,
+        )
+        .context("Failed to create OpenSSL TLS configuration")?;
+        let sigul_client = siguldry::client::Client::new(
+            tls_config,
+            config.sigul.bridge_hostname.clone(),
+            config.sigul.bridge_port,
+            config.sigul.server_hostname.clone(),
+            config.sigul.sigul_user_name.clone(),
+        );
+
         Ok(Self {
             runtime_directory,
-            credentials_directory,
             config,
+            sigul_client,
         })
     }
 }