+++ sigul-pesign-bridge/tests/config.rs
@@ -8,27 +8,21 @@ use assert_cmd::cargo::CommandCargoExt;
 
 #[test]
 fn config_with_default_works() -> Result<()> {
+    let mut creds_directory = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    creds_directory.push("../devel/creds/");
     let mut command = Command::cargo_bin("sigul-pesign-bridge")?;
     let output = command
+        .env("CREDENTIALS_DIRECTORY", creds_directory)
         .arg("config")
         .stderr(Stdio::piped())
         .stdout(Stdio::piped())
         .output()?;
     let logs = String::from_utf8_lossy(&output.stdout);
-    println!("cli_stdout: {logs}");
+    println!("cli_stdout: {}", logs);
+    println!("cli_stderr: {}", String::from_utf8_lossy(&output.stderr));
     assert!(output.status.success());
-    assert!(logs.contains(
-        "Current configuration:
-
-sigul_client_config = \"sigul-client-config\"
-request_timeout_secs = 900
-
-[[keys]]
-key_name = \"signing-key\"
-certificate_name = \"codesigning\"
-passphrase_path = \"sigul-signing-key-passphrase\"
-"
-    ));
+    assert!(logs.contains("[sigul]"));
+    assert!(logs.contains("[[keys]]"));
 
     Ok(())
 }