+++ sigul-pesign-bridge/tests/service_checks_permissions.rs
@@ -1,7 +1,10 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) Microsoft Corporation.
 
-use std::process::{Command, Stdio};
+use std::{
+    path::PathBuf,
+    process::{Command, Stdio},
+};
 
 use anyhow::Result;
 use assert_cmd::cargo::CommandCargoExt;
@@ -10,16 +13,16 @@ use nix::sys::stat::Mode;
 #[test]
 fn stops_world_readable() -> Result<()> {
     let socket_dir = tempfile::tempdir()?;
+    let mut creds_directory = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+    creds_directory.push("../devel/creds/");
 
     nix::sys::stat::umask(Mode::empty());
     let mut command = Command::cargo_bin("sigul-pesign-bridge")?;
     let output = command
         .env("SIGUL_PESIGN_BRIDGE_LOG", "trace")
+        .env("RUNTIME_DIRECTORY", socket_dir.path())
+        .env("CREDENTIALS_DIRECTORY", creds_directory)
         .arg("listen")
-        .arg(format!(
-            "--runtime-directory={}",
-            socket_dir.path().display()
-        ))
         .stderr(Stdio::piped())
         .stdout(Stdio::piped())
         .output()?;