+++ sigul-pesign-bridge/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-03-13
+
+### Changed
+
+- The Python sigul client dependency has been replaced with the siguldry library (#26)
+
+### Added
+
+- It's possible to provide the `--credentials-directory` argument to the `config`
+  sub-command to validate the files exist (#26).
+
+
 ## [0.2.1] - 2025-01-31
 
 ### Added